+++ include/rtkJosephBackProjectionImageFilter.hxx
@@ -61,7 +61,7 @@ JosephBackProjectionImageFilter<TInputImage,
 
   const unsigned int               Dimension = TInputImage::ImageDimension;
   typename TInputImage::RegionType buffReg = this->GetInput(1)->GetBufferedRegion();
-  int                              offsets[3];
+  ptrdiff_t                        offsets[3];
   offsets[0] = 1;
   offsets[1] = this->GetInput(0)->GetBufferedRegion().GetSize()[0];
   offsets[2] =
@@ -129,7 +129,7 @@ JosephBackProjectionImageFilter<TInputImage,
   // Go over each pixel of the projection
   typename BoxShape::VectorType stepMM;
   typename BoxShape::PointType  np, fp;
-  for (unsigned int pix = 0; pix < buffReg.GetNumberOfPixels(); pix++, itIn->Next())
+  for (size_t pix = 0; pix < buffReg.GetNumberOfPixels(); pix++, itIn->Next())
   {
     typename InputRegionIterator::PointType  pixelPosition = itIn->GetPixelPosition();
     typename InputRegionIterator::VectorType dirVox = -itIn->GetSourceToPixel();
@@ -174,9 +174,9 @@ JosephBackProjectionImageFilter<TInputImage,
       const CoordinateType maxy = box->GetBoxMax()[notMainDirSup];
 
       // Init data pointers to first pixel of slice ns (i)nferior and (s)uperior (x|y) corner
-      const int offsetx = offsets[notMainDirInf];
-      const int offsety = offsets[notMainDirSup];
-      int       offsetz = offsets[mainDir];
+      const ptrdiff_t offsetx = offsets[notMainDirInf];
+      const ptrdiff_t offsety = offsets[notMainDirSup];
+      ptrdiff_t       offsetz = offsets[mainDir];
 
       OutputPixelType * pxiyi = beginBuffer + ns * offsetz;
       OutputPixelType * pxsyi = pxiyi + offsetx;
@@ -341,12 +341,12 @@ JosephBackProjectionImageFilter<TInputImage,
                                                              OutputPixelType *      pxsys,
                                                              const double           x,
                                                              const double           y,
-                                                             const int              ox,
-                                                             const int              oy)
+                                                             const ptrdiff_t        ox,
+                                                             const ptrdiff_t        oy)
 {
-  int            ix = itk::Math::floor(x);
-  int            iy = itk::Math::floor(y);
-  int            idx = ix * ox + iy * oy;
+  ptrdiff_t      ix = itk::Math::floor(x);
+  ptrdiff_t      iy = itk::Math::floor(y);
+  ptrdiff_t      idx = ix * ox + iy * oy;
   CoordinateType lx = x - ix;
   CoordinateType ly = y - iy;
   CoordinateType lxc = 1. - lx;
@@ -377,25 +377,25 @@ JosephBackProjectionImageFilter<TInputImage,
                                                                       OutputPixelType *      pxsys,
                                                                       const double           x,
                                                                       const double           y,
-                                                                      const int              ox,
-                                                                      const int              oy,
+                                                                      const ptrdiff_t        ox,
+                                                                      const ptrdiff_t        oy,
                                                                       const CoordinateType   minx,
                                                                       const CoordinateType   miny,
                                                                       const CoordinateType   maxx,
                                                                       const CoordinateType   maxy)
 {
-  int            ix = itk::Math::floor(x);
-  int            iy = itk::Math::floor(y);
-  int            idx = ix * ox + iy * oy;
+  ptrdiff_t      ix = itk::Math::floor(x);
+  ptrdiff_t      iy = itk::Math::floor(y);
+  ptrdiff_t      idx = ix * ox + iy * oy;
   CoordinateType lx = x - ix;
   CoordinateType ly = y - iy;
   CoordinateType lxc = 1. - lx;
   CoordinateType lyc = 1. - ly;
 
-  int offset_xi = 0;
-  int offset_yi = 0;
-  int offset_xs = 0;
-  int offset_ys = 0;
+  ptrdiff_t offset_xi = 0;
+  ptrdiff_t offset_yi = 0;
+  ptrdiff_t offset_xs = 0;
+  ptrdiff_t offset_ys = 0;
 
   if (ix < minx)
     offset_xi = ox;
@@ -433,12 +433,12 @@ JosephBackProjectionImageFilter<TInputImage,
                                                                      const InputPixelType * pxsys,
                                                                      const CoordinateType   x,
                                                                      const CoordinateType   y,
-                                                                     const int              ox,
-                                                                     const int              oy)
+                                                                     const ptrdiff_t        ox,
+                                                                     const ptrdiff_t        oy)
 {
-  int            ix = itk::Math::floor(x);
-  int            iy = itk::Math::floor(y);
-  int            idx = ix * ox + iy * oy;
+  ptrdiff_t      ix = itk::Math::floor(x);
+  ptrdiff_t      iy = itk::Math::floor(y);
+  ptrdiff_t      idx = ix * ox + iy * oy;
   CoordinateType lx = x - ix;
   CoordinateType ly = y - iy;
   CoordinateType lxc = 1. - lx;
@@ -470,25 +470,25 @@ JosephBackProjectionImageFilter<TInputImage,
                                                                               const InputPixelType * pxsys,
                                                                               const CoordinateType   x,
                                                                               const CoordinateType   y,
-                                                                              const int              ox,
-                                                                              const int              oy,
+                                                                              const ptrdiff_t        ox,
+                                                                              const ptrdiff_t        oy,
                                                                               const CoordinateType   minx,
                                                                               const CoordinateType   miny,
                                                                               const CoordinateType   maxx,
                                                                               const CoordinateType   maxy)
 {
-  int            ix = itk::Math::floor(x);
-  int            iy = itk::Math::floor(y);
-  int            idx = ix * ox + iy * oy;
+  ptrdiff_t      ix = itk::Math::floor(x);
+  ptrdiff_t      iy = itk::Math::floor(y);
+  ptrdiff_t      idx = ix * ox + iy * oy;
   CoordinateType lx = x - ix;
   CoordinateType ly = y - iy;
   CoordinateType lxc = 1. - lx;
   CoordinateType lyc = 1. - ly;
 
-  int offset_xi = 0;
-  int offset_yi = 0;
-  int offset_xs = 0;
-  int offset_ys = 0;
+  ptrdiff_t offset_xi = 0;
+  ptrdiff_t offset_yi = 0;
+  ptrdiff_t offset_xs = 0;
+  ptrdiff_t offset_ys = 0;
 
   OutputPixelType result = itk::NumericTraits<typename TOutputImage::PixelType>::ZeroValue();
   if (ix < minx)
